factor out symlink code into symlink.rs
